// src/generated/models/TasksModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface ProjectNameValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface Project_x0020_Name_x003a__x0020_Value {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface AuthorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface EditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface TasksBase {
  ID?: number;
  Title?: string;
  TaskDescription?: string;
  AssignedTo?: string;
  DueDate?: string;
  "ProjectName#Id"?: number;
  "Project_x0020_Name_x003a__x0020_#Id"?: number;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  Modified?: string;
  Created?: string;
  "Author#Claims"?: string;
  "Editor#Claims"?: string;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: {
    "@odata.type"?: string;
    // URL to the small, highly compressed size thumbnail of the item, if available.
    Small?: string;
    // URL to the standard size thumbnail of the item, if available.
    Medium?: string;
    // URL to the largest size thumbnail of the item, if available.
    Large?: string;
  };
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: {
    "@odata.type"?: string;
    Id?: string;
    Name?: string;
  };
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}

export interface TasksRead extends TasksBase {
  ProjectName?: ProjectNameValue;
  Project_x0020_Name_x003a__x0020_?: Project_x0020_Name_x003a__x0020_Value;
  Author?: AuthorValue;
  Editor?: EditorValue;
}

export interface TasksWrite extends TasksBase {
  ProjectName?: ProjectNameValue;
  Project_x0020_Name_x003a__x0020_?: Project_x0020_Name_x003a__x0020_Value;
  Author?: AuthorValue;
  Editor?: EditorValue;
}